Relative CPU Performance Data

Relative CPU Performance Data, described in terms of its cycle time, memory size, etc.

cpu

Format

A data frame with 209 observations on the following 10 variables.

vendor: 30 (adviser, amdahl,apollo, basf, bti, burroughs, c.r.d, cambex, cdc, dec, dg, formation, four-phase, gould, honeywell, hp, ibm, ipl, magnuson, microdata, nas, ncr, nixdorf, perkin-elmer, prime, siemens, sperry, sratus, wang)

Model Name: many unique symbols

MYCT: machine cycle time in nanoseconds (integer)

MMIN: minimum main memory in kilobytes (integer)

MMAX: maximum main memory in kilobytes (integer)

CACH: cache memory in kilobytes (integer)

CHMIN: minimum channels in units (integer)

CHMAX: maximum channels in units (integer)

PRP: published relative performance (integer)

ERP: estimated relative performance from the original article (integer)

Source

Creators: Phillip Ein-Dor and Jacob Feldmesser Ein-Dor: Faculty of Management; Tel Aviv University; Ramat-Aviv; Tel Aviv, 69978; Israel

Donor: David W. Aha (aha@ics.uci.edu) (714) 856-8779

Date: October, 1987

Details

Relative CPU Performance Data, described in terms of its cycle time, memory size, etc. The estimated relative performance values were estimated by the authors using a linear regression method.

References

Ein-Dor and Feldmesser (CACM 4/87, pp 308-317)

Kibler,D. & Aha,D. (1988). Instance-Based Prediction of Real-Valued Attributes. In Proceedings of the CSCSI (Canadian AI) Conference.

https://archive.ics.uci.edu/ml/machine-learning-databases/cpu-performance/

https://archive.ics.uci.edu/ml/datasets/Computer+Hardware