Travel Review Ratings

This data set is populated by capturing user ratings from Google reviews. Reviews on attractions from 24 categories across Europe are considered. Google user rating ranges from 1 to 5 and average user rating per category is calculated.

travel_review_ratings

Format

A data frame with 5456 observations on the following 25 variables.

user id

churches

resorts

beaches

parks

theatres

museums

malls

zoo

restaurants

pubs/bars

local services

burger/pizza shops

hotels/other lodgings

juice bars

art galleries

dance clubs

swimming pools

gyms

bakeries

beauty & spas

cafes

view points

monuments

gardens

Source

Shini Renjith, shinirenjith '@' gmail.com

Details

Google reviews on attractions from 24 categories across Europe are considered. Google user rating ranges from 1 to 5 and average user rating per category is calculated.

References

Renjith, Shini, A. Sreekumar, and M. Jathavedan. 2018. Evaluation of Partitioning Clustering Algorithms for Processing Social Media Data in Tourism Domaina.

https://archive.ics.uci.edu/ml/machine-learning-databases/00485/

https://archive.ics.uci.edu/ml/datasets/Tarvel+Review+Ratings