Zoo

A simple database containing 17 Boolean-valued attributes. The "type" attribute appears to be the class attribute. Here is a breakdown of which animals are in which type: (I find it unusual that there are 2 instances of "frog" and one of "girl"!)

zoo

Format

A data frame with 101 observations on the following 18 variables.

animal name: Unique for each instance

hair: Boolean

feathers: Boolean

eggs: Boolean

milk: Boolean

airborne: Boolean

aquatic: Boolean

predator: Boolean

toothed: Boolean

backbone: Boolean

breathes: Boolean

venomous: Boolean

fins: Boolean

legs: Numeric (set of values: 0,2,4,5,6,8)

tail: Boolean

domestic: Boolean

catsize: Boolean

type: Numeric (integer values in range [1,7], see details)

Source

Creator: Richard Forsyth

Donor: Richard S. Forsyth 8 Grosvenor Avenue Mapperley Park Nottingham NG3 5DX 0602-621676

Date: 5/15/1990

Details

Type Codes:

(41) aardvark, antelope, bear, boar, buffalo, calf, cavy, cheetah, deer, dolphin, elephant, fruitbat, giraffe, girl, goat, gorilla, hamster, hare, leopard, lion, lynx, mink, mole, mongoose, opossum, oryx, platypus, polecat, pony, porpoise, puma, pussycat, raccoon, reindeer, seal, sealion, squirrel, vampire, vole, wallaby,wolf

(20) chicken, crow, dove, duck, flamingo, gull, hawk, kiwi, lark, ostrich, parakeet, penguin, pheasant, rhea, skimmer, skua, sparrow, swan, vulture, wren

(5) pitviper, seasnake, slowworm, tortoise, tuatara

(13) bass, carp, catfish, chub, dogfish, haddock, herring, pike, piranha, seahorse, sole, stingray, tuna

(4) frog, frog, newt, toad

(8) flea, gnat, honeybee, housefly, ladybird, moth, termite, wasp

(10) clam, crab, crayfish, lobster, octopus, scorpion, seawasp, slug, starfish, worm

References

Past Usage: None known other than what is shown in Forsyth's PC/BEAGLE User's Guide.

https://archive.ics.uci.edu/ml/machine-learning-databases/zoo/

https://archive.ics.uci.edu/ml/datasets/Zoo